rest: move rest infra into a module
