+++ crates/sui-e2e-tests/tests/rest/objects.rs
@@ -37,7 +37,10 @@ async fn get_object() {
         // Make sure list works with json
         let _object = client
             .get(url)
-            .header(reqwest::header::ACCEPT, sui_rest_api::APPLICATION_JSON)
+            .header(
+                reqwest::header::ACCEPT,
+                sui_rest_api::rest::APPLICATION_JSON,
+            )
             .send()
             .await
             .unwrap()
@@ -48,7 +51,10 @@ async fn get_object() {
         // Make sure it works with protobuf
         let bytes = client
             .get(url)
-            .header(reqwest::header::ACCEPT, sui_rest_api::APPLICATION_PROTOBUF)
+            .header(
+                reqwest::header::ACCEPT,
+                sui_rest_api::rest::APPLICATION_PROTOBUF,
+            )
             .send()
             .await
             .unwrap()
@@ -61,7 +67,7 @@ async fn get_object() {
         // protobuf version
         let bytes = client
             .get(url)
-            .header(reqwest::header::ACCEPT, sui_rest_api::APPLICATION_BCS)
+            .header(reqwest::header::ACCEPT, sui_rest_api::rest::APPLICATION_BCS)
             .send()
             .await
             .unwrap()
