+++ crates/sui-e2e-tests/tests/rest/resolve.rs
@@ -5,7 +5,7 @@ use shared_crypto::intent::Intent;
 use sui_keys::keystore::AccountKeystore;
 use sui_macros::sim_test;
 use sui_rest_api::client::reqwest::StatusCode;
-use sui_rest_api::transactions::ResolveTransactionQueryParameters;
+use sui_rest_api::rest::transactions::ResolveTransactionQueryParameters;
 use sui_rest_api::Client;
 use sui_rest_api::ExecuteTransactionQueryParameters;
 use sui_sdk_types::types::unresolved;
