+++ crates/sui-rest-api/src/lib.rs
@@ -1,134 +1,33 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use axum::{
-    response::{Redirect, ResponseParts},
-    routing::get,
-    Router,
-};
 use mysten_network::callback::CallbackLayer;
-use openapi::ApiEndpoint;
 use reader::StateReader;
+use rest::build_rest_router;
 use std::sync::Arc;
 use sui_types::storage::RpcStateReader;
 use sui_types::transaction_executor::TransactionExecutor;
 use tap::Pipe;
 
-pub mod accept;
-mod accounts;
-mod checkpoints;
 pub mod client;
-mod coins;
-mod committee;
 mod config;
-pub mod content_type;
 mod error;
-mod health;
-mod info;
 mod metrics;
-mod objects;
-pub mod openapi;
 pub mod proto;
 mod reader;
 mod response;
-mod system;
-pub mod transactions;
+pub mod rest;
 pub mod types;
 
-pub use checkpoints::CheckpointResponse;
-pub use checkpoints::ListCheckpointsQueryParameters;
 pub use client::Client;
+pub use config::Config;
 pub use error::{RestError, Result};
 pub use metrics::RestMetrics;
-pub use objects::ObjectResponse;
+pub use rest::checkpoints::CheckpointResponse;
+pub use rest::checkpoints::ListCheckpointsQueryParameters;
+pub use rest::objects::ObjectResponse;
+pub use rest::transactions::ExecuteTransactionQueryParameters;
 pub use sui_types::full_checkpoint_content::{CheckpointData, CheckpointTransaction};
-pub use transactions::ExecuteTransactionQueryParameters;
-pub use config::Config;
-
-pub const TEXT_PLAIN_UTF_8: &str = "text/plain; charset=utf-8";
-pub const APPLICATION_BCS: &str = "application/bcs";
-pub const APPLICATION_JSON: &str = "application/json";
-pub const APPLICATION_PROTOBUF: &str = "application/x-protobuf";
-
-#[derive(Debug, Copy, Clone, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
-#[serde(rename_all = "lowercase")]
-pub enum Direction {
-    Ascending,
-    Descending,
-}
-
-impl Direction {
-    pub fn is_descending(self) -> bool {
-        matches!(self, Self::Descending)
-    }
-}
-
-#[derive(Debug)]
-pub struct Page<T, C> {
-    pub entries: response::ResponseContent<Vec<T>>,
-    pub cursor: Option<C>,
-}
-
-pub struct PageCursor<C>(pub Option<C>);
-
-impl<C: std::fmt::Display> axum::response::IntoResponseParts for PageCursor<C> {
-    type Error = (axum::http::StatusCode, String);
-
-    fn into_response_parts(
-        self,
-        res: ResponseParts,
-    ) -> std::result::Result<ResponseParts, Self::Error> {
-        self.0
-            .map(|cursor| [(crate::types::X_SUI_CURSOR, cursor.to_string())])
-            .into_response_parts(res)
-            .map_err(|e| (axum::http::StatusCode::INTERNAL_SERVER_ERROR, e.to_string()))
-    }
-}
-
-impl<C: std::fmt::Display> axum::response::IntoResponse for PageCursor<C> {
-    fn into_response(self) -> axum::response::Response {
-        (self, ()).into_response()
-    }
-}
-
-pub const DEFAULT_PAGE_SIZE: usize = 50;
-pub const MAX_PAGE_SIZE: usize = 100;
-
-impl<T: serde::Serialize, C: std::fmt::Display> axum::response::IntoResponse for Page<T, C> {
-    fn into_response(self) -> axum::response::Response {
-        let cursor = self
-            .cursor
-            .map(|cursor| [(crate::types::X_SUI_CURSOR, cursor.to_string())]);
-
-        (cursor, self.entries).into_response()
-    }
-}
-
-pub const ENDPOINTS: &[&dyn ApiEndpoint<RestService>] = &[
-    // stable APIs
-    &info::GetNodeInfo,
-    &health::HealthCheck,
-    &checkpoints::ListCheckpoints,
-    &checkpoints::GetCheckpoint,
-    // unstable APIs
-    &accounts::ListAccountObjects,
-    &objects::GetObject,
-    &objects::GetObjectWithVersion,
-    &objects::ListDynamicFields,
-    &checkpoints::GetFullCheckpoint,
-    &transactions::GetTransaction,
-    &transactions::ListTransactions,
-    &committee::GetCommittee,
-    &committee::GetLatestCommittee,
-    &system::GetSystemStateSummary,
-    &system::GetCurrentProtocolConfig,
-    &system::GetProtocolConfig,
-    &system::GetGasInfo,
-    &transactions::ExecuteTransaction,
-    &transactions::SimulateTransaction,
-    &transactions::ResolveTransaction,
-    &coins::GetCoinInfo,
-];
 
 #[derive(Clone)]
 pub struct RestService {
@@ -138,18 +37,6 @@ pub struct RestService {
     software_version: &'static str,
     metrics: Option<Arc<RestMetrics>>,
     config: Config,
-}
-
-impl axum::extract::FromRef<RestService> for StateReader {
-    fn from_ref(input: &RestService) -> Self {
-        input.reader.clone()
-    }
-}
-
-impl axum::extract::FromRef<RestService> for Option<Arc<dyn TransactionExecutor>> {
-    fn from_ref(input: &RestService) -> Self {
-        input.executor.clone()
-    }
 }
 
 impl RestService {
@@ -189,26 +76,12 @@ impl RestService {
         self.software_version
     }
 
-    pub fn into_router(self) -> Router {
+    pub fn into_router(self) -> axum::Router {
         let metrics = self.metrics.clone();
 
-        let mut api = openapi::Api::new(info(self.software_version()));
+        let rest_router = build_rest_router(self.clone());
 
-        api.register_endpoints(
-            ENDPOINTS
-                .iter()
-                .copied()
-                .filter(|endpoint| endpoint.stable() || self.config.enable_unstable_apis()),
-        );
-
-        Router::new()
-            .nest("/v2/", api.to_router().with_state(self.clone()))
-            .route("/v2", get(|| async { Redirect::permanent("/v2/") }))
-            // Previously the service used to be hosted at `/rest`. In an effort to migrate folks
-            // to the new versioned route, we'll issue redirects from `/rest` -> `/v2`.
-            .route("/rest/*path", axum::routing::method_routing::any(redirect))
-            .route("/rest", get(|| async { Redirect::permanent("/v2/") }))
-            .route("/rest/", get(|| async { Redirect::permanent("/v2/") }))
+        rest_router
             .layer(axum::middleware::map_response_with_state(
                 self,
                 response::append_info_headers,
@@ -230,60 +103,15 @@ impl RestService {
     }
 }
 
-pub fn info(version: &'static str) -> openapiv3::v3_1::Info {
-    use openapiv3::v3_1::Contact;
-    use openapiv3::v3_1::License;
-
-    openapiv3::v3_1::Info {
-        title: "Sui Node Api".to_owned(),
-        description: Some("REST Api for interacting with the Sui Blockchain".to_owned()),
-        contact: Some(Contact {
-            name: Some("Mysten Labs".to_owned()),
-            url: Some("https://github.com/MystenLabs/sui".to_owned()),
-            ..Default::default()
-        }),
-        license: Some(License {
-            name: "Apache 2.0".to_owned(),
-            url: Some("https://www.apache.org/licenses/LICENSE-2.0.html".to_owned()),
-            ..Default::default()
-        }),
-        version: version.to_owned(),
-        ..Default::default()
-    }
+#[derive(Debug, Copy, Clone, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+#[serde(rename_all = "lowercase")]
+pub enum Direction {
+    Ascending,
+    Descending,
 }
 
-async fn redirect(axum::extract::Path(path): axum::extract::Path<String>) -> Redirect {
-    Redirect::permanent(&format!("/v2/{path}"))
-}
-
-mod _schemars {
-    use schemars::schema::InstanceType;
-    use schemars::schema::Metadata;
-    use schemars::schema::SchemaObject;
-    use schemars::JsonSchema;
-
-    pub(crate) struct U64;
-
-    impl JsonSchema for U64 {
-        fn schema_name() -> String {
-            "u64".to_owned()
-        }
-
-        fn json_schema(_: &mut schemars::gen::SchemaGenerator) -> schemars::schema::Schema {
-            SchemaObject {
-                metadata: Some(Box::new(Metadata {
-                    description: Some("Radix-10 encoded 64-bit unsigned integer".to_owned()),
-                    ..Default::default()
-                })),
-                instance_type: Some(InstanceType::String.into()),
-                format: Some("u64".to_owned()),
-                ..Default::default()
-            }
-            .into()
-        }
-
-        fn is_referenceable() -> bool {
-            false
-        }
+impl Direction {
+    pub fn is_descending(self) -> bool {
+        matches!(self, Self::Descending)
     }
 }
