+++ crates/sui-rest-api/src/rest/transactions/resolve/mod.rs
@@ -6,16 +6,16 @@ use std::collections::HashMap;
 
 use super::execution::SimulateTransactionQueryParameters;
 use super::TransactionSimulationResponse;
-use crate::accept::AcceptJsonProtobufBcs;
-use crate::objects::ObjectNotFoundError;
-use crate::openapi::ApiEndpoint;
-use crate::openapi::OperationBuilder;
-use crate::openapi::RequestBodyBuilder;
-use crate::openapi::ResponseBuilder;
-use crate::openapi::RouteHandler;
 use crate::proto;
 use crate::reader::StateReader;
 use crate::response::JsonProtobufBcs;
+use crate::rest::accept::AcceptJsonProtobufBcs;
+use crate::rest::objects::ObjectNotFoundError;
+use crate::rest::openapi::ApiEndpoint;
+use crate::rest::openapi::OperationBuilder;
+use crate::rest::openapi::RequestBodyBuilder;
+use crate::rest::openapi::ResponseBuilder;
+use crate::rest::openapi::RouteHandler;
 use crate::RestError;
 use crate::RestService;
 use crate::Result;
